Provide links to the relevant pages for anyone with ideas about how to make /all better
Humans will get annoyed and complain rather than going and doing something.
It's like we don't realize we have the power to make a change unless someone explicitly says so.
Also, even when we do realize this, if the bootstrapping is sufficiently difficult, we'll get derailed.

So, explicitly tell users that this is a thing they can do, and show them where they can go to do it
Bc I'm not going to sit and listen to some someone whine about how something I made better isn't perfect,
if they don't like it, they should go fix it.

This is the invitation with directions to the party, BYOPR.